training.py
This file will train your custom data, You have to give the path of the main folder where all the folders of different people are, with their names as folder names.  each have individual images of that person, This will auto preprocess and resize the image to get the face encodings.